requirements.txt
Flask==0.10.1
Flask-Assets==0.9
Flask-Login==0.2.11
Flask-BasicAuth==0.2.0
Flask-SQLAlchemy==1.0
Flask-Debugtoolbar==0.9.0
Flask-Mail==0.9.0
Flask-Principal==0.4.0
Flask-Security==1.7.3
Flask-WTF==0.9.5
Flask-Bcrypt==0.6
Flask-Migrate==1.2.0
Jinja2==2.7.2
MarkupSafe==0.21
Werkzeug==0.9.4
gunicorn==18.0
requests==2.3.0
SQLAlchemy==0.9.6
itsdangerous==0.24
psycopg2==2.5.3
py-bcrypt==0.4
